# Add the OpenMP optimization for UpsampleNearest and BatchPermutation.