WIth this optimization, the following two ops can boost a lot. (Test with MaskRCNN, on SKX8180 one socket)
UpsampleNearest op: reduced from 240.7196 ms to 9.225 ms
BatchPermutation op: reduced from 8.296387 ms to 1.4501984 ms.