@ChongyuIntel 2018-09-18
This is for Caffe2 optimization.